[mypyc] Refactor mypyc error handling
Pull Errors out into its own module and have it get managed externally
from genops. This is in preparation for separate/incremental
compilation.
