+++ mypyc/test/testutil.py
@@ -17,6 +17,7 @@ from mypy.test.helpers import assert_string_arrays_equal
 from mypyc import genops
 from mypyc.options import CompilerOptions
 from mypyc.ops import FuncIR
+from mypyc.errors import Errors
 from mypyc.test.config import test_data_prefix
 
 # The builtins stub used during icode generation test cases.
@@ -102,9 +103,12 @@ def build_ir_for_single_file(input_lines: List[str],
                          alt_lib_path=test_temp_dir)
     if result.errors:
         raise CompileError(result.errors)
-    _, modules, errors = genops.build_ir([result.files['__main__']], result.graph, result.types,
-                                         compiler_options)
-    assert errors == 0
+
+    errors = Errors()
+    _, modules = genops.build_ir(
+        [result.files['__main__']], result.graph, result.types,
+        compiler_options, errors)
+    assert errors.num_errors == 0
 
     module = modules[0][1]
     return module.functions
